Refine test cases
