+++ mars/tests/core.py
@@ -13,14 +13,14 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import functools
+import itertools
 import os
-import sys
 import logging
 import shutil
 import subprocess
 import tempfile
 import time
-import itertools
 import unittest
 from collections.abc import Iterable
 from weakref import ReferenceType
@@ -42,6 +42,8 @@ _mock = mock
 
 cupy = lazy_import('cupy', globals=globals())
 cudf = lazy_import('cudf', globals=globals())
+
+logger = logging.getLogger(__name__)
 
 
 class TestCase(unittest.TestCase):
@@ -67,13 +69,27 @@ def parameterized(**params):
     the test case's name will generated by dictionary keys.
     :return: Generated test classes
     """
-    def decorator(base_class):
-        module = sys.modules[base_class.__module__].__dict__
-        for n, param in params.items():
-            param_dict = dict(base_class.__dict__)
-            param_dict.update(param)
-            name = "%s%s" % (base_class.__name__, n.capitalize())
-            module[name] = type(name, (base_class,), param_dict)
+    def _build_parameterized_case(cls, fun):
+        @functools.wraps(fun)
+        def test_fun(self, *args, **kwargs):
+            for n, param in params.items():
+                for pn, pv in param.items():
+                    setattr(cls, pn, pv)
+                try:
+                    fun(self, *args, **kwargs)
+                except:  # noqa: E722
+                    logger.error('Failed when running %s with param %s', fun.__name__, n)
+                    raise
+        return test_fun
+
+    def decorator(cls):
+        fun_dict = dict()
+        for fun_name, fun in cls.__dict__.items():
+            if callable(fun) and fun_name.startswith('test'):
+                fun_dict[fun_name] = _build_parameterized_case(cls, fun)
+        for fun_name, fun in fun_dict.items():
+            setattr(cls, fun_name, fun)
+        return cls
 
     return decorator
